requirements.txt
# Core data ML libraries
pandas>=2.3.0
numpy>=2.3.1
matplotlib>=3.10.3
seaborn>=0.13.2

# Deep learning frameworks
torch==2.7.1

# MLX libraries (for Apple Silicon)
mlx>=0.26.1
mlx-lm>=0.25.2

# Reinforcement learning
gymnasium>=1.1.1
stable-baselines3>=2.6.0

# Tokenization
tokenizers>=0.21.2

# Terminal output
rich>=14.0.0

# Configuration
PyYAML>=6.0.2

# Utilities
tqdm>=4.67.1
requests>=2.32.4

# Web3
web3>=7.12.0
eth-abi>=5.2.0
eth-account>=0.13.7
uniswap_universal_router_decoder>=2.0.0